build pass

+++ src/modules/navigator/mission_base.cpp
@@ -73,9 +73,9 @@ MissionBase::updateDatamanCache()
 {
 	if ((_mission.count > 0) && (_mission.current_seq != _load_mission_index)) {
 
-		const int32_t start_index = math::constrain(_mission.current_seq, INT32_C(0), int32_t(_mission.count) - 1);
-		const int32_t end_index = math::constrain(start_index + _dataman_cache_size_signed, INT32_C(0),
-					  int32_t(_mission.count) - 1);
+		const int32_t start_index = math::constrain((int32_t)_mission.current_seq, (int32_t)INT32_C(0), (int32_t)int32_t(_mission.count) - 1);
+		const int32_t end_index = math::constrain((int32_t)start_index + _dataman_cache_size_signed, (int32_t)INT32_C(0),
+					  (int32_t)int32_t(_mission.count) - 1);
 
 		for (int32_t index = start_index; index != end_index; index += math::signNoZero(_dataman_cache_size_signed)) {
 
@@ -98,8 +98,8 @@ void MissionBase::updateMavlinkMission()
 		const bool mission_data_changed = checkMissionDataChanged(new_mission);
 
 		if (new_mission.current_seq < 0) {
-			new_mission.current_seq = math::constrain(_mission.current_seq, INT32_C(0),
-						  static_cast<int32_t>(new_mission.count) - 1);
+			new_mission.current_seq = math::constrain((int32_t)_mission.current_seq, (int32_t)INT32_C(0),
+						  (int32_t)static_cast<int32_t>(new_mission.count) - 1);
 		}
 
 		if (new_mission.geofence_id != _mission.geofence_id) {
